# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: EcoStride API
  version: 0.1.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /admin/sponsors:
    get:
      summary: List sponsors
      tags: [Admin Sponsors]
      responses:
        "200": { description: OK }
    post:
      summary: Create sponsor
      tags: [Admin Sponsors]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "201": { description: Created }
  /admin/sponsors/{id}:
    put:
      summary: Update sponsor
      tags: [Admin Sponsors]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
    delete:
      summary: Delete sponsor
      tags: [Admin Sponsors]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...